refactor: 폴더 구조 변경 및 guestInsurance 라우팅 수정 - atoms 폴더 추가 - /insurance/guest 로 미가입자 보험 페이지 라우팅 수정
