+++ src/pages/Insurance/GuestInsurance.module.scss
@@ -1,8 +1,8 @@
 @import 'src/styles/style.scss';
 
 .container {
-  position: absolute;
-  background-color: $background;
+  width: 100%;
+  margin-bottom: 30px;
 }
 
 .subHeader {
@@ -114,7 +114,6 @@
   height: 477px;
   margin-top: 24px;
   margin-left: 41px;
-  margin-bottom: 30px;
   padding: 40px 25px 38px 25px;
 
   background-color: $white;
