+++ src/pages/Insurance/atoms/ProgressBar/ProgressBar.module.scss
